draftexamples/influx/utils/torqueutils/requirements.txt
lxml==4.6.2
beautifulsoup4==4.9.3
PyYAML==5.4.1
rich==9.11.1
